requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
twilio==9.3.3
elevenlabs==1.7.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
pydantic==2.9.2
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
